requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
jinja2==3.1.2
python-multipart==0.0.6
transformers==4.35.2
torch==2.2.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pydantic==2.5.0
setuptools>=65.0.0
wheel>=0.38.0 